move main.py into Objects directory, and rename it to XRule.py
