+++ quark/cli.py
@@ -5,7 +5,7 @@ from tqdm import tqdm
 
 from quark.Objects.RuleObject import RuleObject
 from quark.logo import logo
-from quark.main import XRule
+from quark.Objects.XRule import XRule
 from quark.utils.out import print_success, print_info, print_warning
 from quark.utils.weight import Weight
