+++ tests/test_main.py
@@ -1,1 +1,1 @@
-from quark.main import XRule
+
